[fix] : parsing structure modifying
